refact: restructure of actions and functions, action accepts type ref temporary
