+++ src/ast/cdl/actions_and_functions/mod.rs
@@ -1,5 +1,11 @@
 mod action_declaration_term;
 mod action_term;
+mod function_term;
+mod param_term;
+mod returns_term;
 
 pub use action_declaration_term::*;
 pub use action_term::*;
+pub use function_term::*;
+pub use param_term::*;
+pub use returns_term::*;
