+++ src/ast/cdl/actions_and_functions/returns_term.rs
@@ -1,4 +1,4 @@
-use super::name_term::NameTerm;
+use super::super::name_term::NameTerm;
 use ast_term_derive::ASTTerm;
 
 #[derive(ASTTerm, PartialEq, Eq, Debug, Clone)]
