+++ src/ast/cdl/mod.rs
@@ -2,12 +2,9 @@ mod actions_and_functions;
 mod entity_term;
 mod field_term;
 mod from_term;
-mod function_term;
 mod module_term;
 mod name_term;
-mod param_term;
 mod path_term;
-mod returns_term;
 mod service_term;
 mod types;
 
@@ -15,11 +12,8 @@ pub use actions_and_functions::*;
 pub use entity_term::*;
 pub use field_term::*;
 pub use from_term::*;
-pub use function_term::*;
 pub use module_term::*;
 pub use name_term::*;
-pub use param_term::*;
 pub use path_term::*;
-pub use returns_term::*;
 pub use service_term::*;
 pub use types::*;
