+++ src/ast/cdl/service_term.rs
@@ -1,7 +1,7 @@
 use crate::ast::*;
 
 use super::entity_term::EntityTerm;
-use super::function_term::FunctionTerm;
+use super::FunctionTerm;
 use super::name_term::NameTerm;
 use super::ActionDeclarationTerm;
 use super::TypeDeclarationTerm;
